requirements.txt
autopep8==1.4.4
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
lockfile==0.12.2
luigi==2.8.3
numpy==1.16.2
oauthlib==3.0.1
opencv-python==4.1.0.25
pandas==0.24.2
pathlib==1.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
PySocks==1.6.8
python-daemon==2.1.2
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.3
tornado==4.5.3
tweepy==3.7.0
uritemplate==3.0.0
urllib3==1.24.1
yagmail==0.11.215
urllib3==1.24.2